prevent false positive cyclic dependency when root directory name clashes with a dependency name
This detects incorrect overrides by validating that the overriden package fulfils any product dependencies that the original would fulfill and if that is not the case, will error about an incorrect override during dependency resolution. This can only work if the client of the package uses tools version 5.2 or later since otherwise we cannot be certain that the product dependency should be fulfillable.